Edit Navigation Component
