+++ app/build.gradle
@@ -38,6 +38,7 @@ dependencies {
     def nav_version = "2.3.0-alpha04"
     implementation "androidx.navigation:navigation-fragment:$nav_version"
     implementation "androidx.navigation:navigation-ui:$nav_version"
+    implementation 'com.github.xiangtailiang:tablayout:1.0.0'
 
 
 }
