+++ app/src/main/java/com/engineering/dokkan/view/Favourite/ViewPagerAdapter.java
@@ -17,6 +17,7 @@ public class ViewPagerAdapter extends FragmentPagerAdapter {
 
     public ViewPagerAdapter(FragmentManager fm) {
         super(fm, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT);
+        //super(fm);
 
     }
