+++ app/src/main/java/com/engineering/dokkan/view/MainActivity.java
@@ -24,7 +24,7 @@ public class MainActivity extends AppCompatActivity {
 //        navigateToProfile();
 //        navigateToNotification();
 //        navigateToChat();
-          navigateToQuestions();
+   //       navigateToQuestions();
 
 
     }
