+++ app/src/main/res/layout/fragment_favourite.xml
@@ -22,6 +22,7 @@
         app:tabTextColor="@color/colorAccent">
 
         <com.google.android.material.tabs.TabItem
+            android:id="@+id/ItemTab"
             android:layout_width="@dimen/width_5"
             android:layout_height="wrap_content"
             android:layout_gravity="center"
@@ -29,6 +30,7 @@
 
 
         <com.google.android.material.tabs.TabItem
+            android:id="@+id/shopTab"
             android:layout_width="@dimen/width_5"
             android:layout_height="wrap_content"
             android:layout_gravity="center"
