+++ app/src/main/res/layout/recycler_favourite_item.xml
@@ -51,7 +51,9 @@
             android:paddingLeft="@dimen/paddingLeft_4"
             android:text="Full Item Name"
             android:textSize="@dimen/textsize_10"
-            android:layout_below="@+id/itemImage"/>
+            android:layout_below="@+id/itemImage"
+            android:textColor="@color/txtColor_black"
+            />
 
 
            <RatingBar
